requirements.txt
asgiref==3.8.0
Django==5.0.3
djangorestframework==3.15.0
dotenv-python==0.0.1
python-dotenv==1.0.1
sqlparse==0.4.4
asgiref==3.8.0
Django==5.0.3
djangorestframework==3.15.0
dotenv-python==0.0.1
pillow==10.2.0
python-dotenv==1.0.1
sqlparse==0.4.4
